requirements.txt
typer>=0.9.0
rich>=13.7.0
PyYAML>=6.0.1
pydantic>=2.6.0
datasets>=2.19.0
pyarrow>=14.0.0
Pillow>=10.0.0
transformers>=4.40.0
tokenizers>=0.15.0
jinja2